+++ tket/test/src/Passes/test_CliffordResynthesis.cpp
@@ -92,13 +92,21 @@ SCENARIO("SynthesiseCliffordResynthesis correctness") {
     c.add_op<unsigned>(OpType::ZZMax, {0, 1});
     check_clifford_resynthesis(c);
   }
-  GIVEN("A troublesome circuit") {
+  GIVEN("A troublesome circuit (1)") {
     // https://github.com/CQCL/tket/issues/1279
     Circuit c(3);
     c.add_op<unsigned>(OpType::ECR, {1, 2});
     c.add_op<unsigned>(OpType::CnRy, 0., {0, 1});
     c.add_op<unsigned>(OpType::Rz, 0., {1});
     c.add_op<unsigned>(OpType::ZZMax, {2, 1});
+    check_clifford_resynthesis(c);
+  }
+  GIVEN("A troublesome circuit (2)") {
+    Circuit c(2);
+    c.add_op<unsigned>(OpType::CX, {0, 1});
+    c.add_op<unsigned>(OpType::Z, {0});
+    c.add_op<unsigned>(OpType::T, {0});
+    c.add_op<unsigned>(OpType::CY, {1, 0});
     check_clifford_resynthesis(c);
   }
 }